+++ src/extension.ts
@@ -2,6 +2,8 @@ import * as vscode from 'vscode';
 import fetch from "node-fetch";
 
 const SalimAPIUrl: string = "https://watasalim.vercel.app/api/quotes"
+const AllSalimQuotes: string = "https://github.com/narze/awesome-salim-quotes/blob/main/README.md"
+
 const LinkToHeaven: string = "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
 const RickRollChance: number = 5
 
@@ -22,20 +24,31 @@ export async function activate(context: vscode.ExtensionContext) {
         })
         .catch(err => console.log(err))
 
-    let disposable: vscode.Disposable = vscode.commands.registerCommand('salim-quote-delivery.getSalimQuote', () => {
+    let deliwry: vscode.Disposable = vscode.commands.registerCommand('salim-quote-delivery.getSalimQuote', () => {
+        // * Check for Errors
+        if (QuoteArray.length == 0) {
+            vscode.window.showErrorMessage("ERROR: Quote Array is empty. There maybe issues initializing or API is down.")
+            return
+        }
+
         // * Random Quote Index
         let randIndex: number = Math.floor(Math.random() * QuoteArray.length)
 
         // * Show it
-        vscode.window.showInformationMessage(`${QuoteArray[randIndex]}`);
+        vscode.window.showInformationMessage(`${QuoteArray[randIndex]} (#${randIndex + 1})`);
 
         // * You have a Chance to be sent to Heaven
         if (Math.floor(Math.random() * 100) < RickRollChance) {
             vscode.env.openExternal(vscode.Uri.parse(LinkToHeaven))
         }
-    });
+    })
 
-    context.subscriptions.push(disposable);
+    let showAllQuotes: vscode.Disposable = vscode.commands.registerCommand('salim-quote-delivery.showAllQuotes', () => {
+        vscode.env.openExternal(vscode.Uri.parse(AllSalimQuotes))
+    })
+
+    context.subscriptions.push(deliwry)
+    context.subscriptions.push(showAllQuotes)
 }
 
 export function deactivate() { }